pyproject.toml
[project]
name = "wire-upgrade"
version = "0.1.0"
description = "Wire Upgrade CLI"
requires-python = ">=3.10"
dependencies = [
  "typer>=0.9.0",
  "rich>=13.0.0",
  "pydantic>=2.0.0",
  "PyYAML>=6.0",
]

[project.scripts]
wire-upgrade = "wire_upgrade.orchestrator:app"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["wire_upgrade"]
